+++ lib/exception_handling/exception_description.rb
@@ -6,12 +6,13 @@ module ExceptionHandling
 
     CONFIGURATION_SECTIONS = {
       send_to_honeybadger: false,  # should be sent to honeybadger?
+      send_to_sentry: false,       # should be sent to sentry?
       send_metric: true,   # should the metric be sent.
       metric_name: nil,    # Will be derived from section name if not passed
       notes: nil     # Will be included in exception email if set, used to keep notes and relevant links
     }.freeze
 
-    attr_reader :filter_name, :send_to_honeybadger, :send_metric, :metric_name, :notes
+    attr_reader :filter_name, :send_to_honeybadger, :send_to_sentry, :send_metric, :metric_name, :notes
 
     def initialize(filter_name, configuration)
       @filter_name = filter_name
@@ -21,6 +22,7 @@ module ExceptionHandling
 
       @configuration = CONFIGURATION_SECTIONS.merge(configuration)
       @send_to_honeybadger = @configuration[:send_to_honeybadger]
+      @send_to_sentry = @configuration[:send_to_sentry]
       @send_metric = @configuration[:send_metric]
       @metric_name = (@configuration[:metric_name] || @filter_name).to_s.gsub(" ", "_")
       @notes       = @configuration[:notes]